Fix relay leaderboard
A single block can be relayed by multiple relays. Presumably this is for
resilience in case of a relay going down. In the past we only stored one
relay per slot, now we store all of them. Misses and market shares are
divided between these relays, so e.g. if a block with 2 misses is
relayed by 3 relays, each relay is attributed 2/3 misses.